requirements.txt
asgiref==3.10.0
black==25.11.0
cfgv==3.4.0
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
inflection==0.5.1
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pre_commit==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.35.4
